Fix the scenario test gates
We currently use a gate hook that adds
a hardcoded image name which is cirros-0.3.5-x86_64-disk to work
around a devstack bug[2].

A recent patch[1] changed the deployed image to
cirros-0.3.6-x86_64-disk.

Now that the bug[2] has been resolved we no longer need to pin the
cirros image config in our gates.

[1] https://review.openstack.org/624716
[2] https://review.openstack.org/#/c/435106

Change-Id: Ia41577b38eca94617075f849a88ff3c126e407eb