remove release notes job
Since the master branch of neutron-lbaas has been removed (retired),
the release notes job will not longer work as it runs release notes
build from master branch which does not exist any more for lbaas.

The infra/release team need to discuss how to address this issue, but
for now this patch proposes to just remove the release notes job. We
can add it back if/when a conclusion is reached on how to address it.

Change-Id: Ib9b1ee5af2a7a9d2b9838a4bbd8e9b8670749886